Update Game Controller to ignore inputs while solving
The Game Controller now ignores any mouse clicks if the game is in 'solving' state. This prevents unexpected behavior during automatic solving process. Also, excessive whitespaces were removed and comments have been added for better readability and understanding of the code. SDL2_image package was removed from CMakeLists as it was not in use.